requirement.txt
anyio==4.0.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
comm==0.1.4
contourpy==1.1.0
cvzone==1.5.6
cycler==0.11.0
Cython==3.0.0
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.12.2
fonttools==4.41.0
fpdf2==2.7.5
fqdn==1.5.1
idna==3.4
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keyboard==0.13.5
kiwisolver==1.4.4
lapx==0.5.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
memory-profiler==0.61.0
mistune==3.0.1
mpmath==1.3.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.25.1
opencv-python==4.8.0.74
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycparser==2.21
Pygments==2.16.1
pyobjc==9.2
pyobjc-core==9.2
pyobjc-framework-Accessibility==9.2
pyobjc-framework-Accounts==9.2
pyobjc-framework-AddressBook==9.2
pyobjc-framework-AdServices==9.2
pyobjc-framework-AdSupport==9.2
pyobjc-framework-AppleScriptKit==9.2
pyobjc-framework-AppleScriptObjC==9.2
pyobjc-framework-ApplicationServices==9.2
pyobjc-framework-AppTrackingTransparency==9.2
pyobjc-framework-AudioVideoBridging==9.2
pyobjc-framework-AuthenticationServices==9.2
pyobjc-framework-AutomaticAssessmentConfiguration==9.2
pyobjc-framework-Automator==9.2
pyobjc-framework-AVFoundation==9.2
pyobjc-framework-AVKit==9.2
pyobjc-framework-AVRouting==9.2
pyobjc-framework-BackgroundAssets==9.2
pyobjc-framework-BusinessChat==9.2
pyobjc-framework-CalendarStore==9.2
pyobjc-framework-CallKit==9.2
pyobjc-framework-CFNetwork==9.2
pyobjc-framework-ClassKit==9.2
pyobjc-framework-CloudKit==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-Collaboration==9.2
pyobjc-framework-ColorSync==9.2
pyobjc-framework-Contacts==9.2
pyobjc-framework-ContactsUI==9.2
pyobjc-framework-CoreAudio==9.2
pyobjc-framework-CoreAudioKit==9.2
pyobjc-framework-CoreBluetooth==9.2
pyobjc-framework-CoreData==9.2
pyobjc-framework-CoreHaptics==9.2
pyobjc-framework-CoreLocation==9.2
pyobjc-framework-CoreMedia==9.2
pyobjc-framework-CoreMediaIO==9.2
pyobjc-framework-CoreMIDI==9.2
pyobjc-framework-CoreML==9.2
pyobjc-framework-CoreMotion==9.2
pyobjc-framework-CoreServices==9.2
pyobjc-framework-CoreSpotlight==9.2
pyobjc-framework-CoreText==9.2
pyobjc-framework-CoreWLAN==9.2
pyobjc-framework-CryptoTokenKit==9.2
pyobjc-framework-DataDetection==9.2
pyobjc-framework-DeviceCheck==9.2
pyobjc-framework-DictionaryServices==9.2
pyobjc-framework-DiscRecording==9.2
pyobjc-framework-DiscRecordingUI==9.2
pyobjc-framework-DiskArbitration==9.2
pyobjc-framework-DVDPlayback==9.2
pyobjc-framework-EventKit==9.2
pyobjc-framework-ExceptionHandling==9.2
pyobjc-framework-ExecutionPolicy==9.2
pyobjc-framework-ExtensionKit==9.2
pyobjc-framework-ExternalAccessory==9.2
pyobjc-framework-FileProvider==9.2
pyobjc-framework-FileProviderUI==9.2
pyobjc-framework-FinderSync==9.2
pyobjc-framework-FSEvents==9.2
pyobjc-framework-GameCenter==9.2
pyobjc-framework-GameController==9.2
pyobjc-framework-GameKit==9.2
pyobjc-framework-GameplayKit==9.2
pyobjc-framework-HealthKit==9.2
pyobjc-framework-ImageCaptureCore==9.2
pyobjc-framework-IMServicePlugIn==9.2
pyobjc-framework-InputMethodKit==9.2
pyobjc-framework-InstallerPlugins==9.2
pyobjc-framework-InstantMessage==9.2
pyobjc-framework-Intents==9.2
pyobjc-framework-IntentsUI==9.2
pyobjc-framework-IOBluetooth==9.2
pyobjc-framework-IOBluetoothUI==9.2
pyobjc-framework-IOSurface==9.2
pyobjc-framework-iTunesLibrary==9.2
pyobjc-framework-KernelManagement==9.2
pyobjc-framework-LatentSemanticMapping==9.2
pyobjc-framework-LaunchServices==9.2
pyobjc-framework-libdispatch==9.2
pyobjc-framework-libxpc==9.2
pyobjc-framework-LinkPresentation==9.2
pyobjc-framework-LocalAuthentication==9.2
pyobjc-framework-LocalAuthenticationEmbeddedUI==9.2
pyobjc-framework-MailKit==9.2
pyobjc-framework-MapKit==9.2
pyobjc-framework-MediaAccessibility==9.2
pyobjc-framework-MediaLibrary==9.2
pyobjc-framework-MediaPlayer==9.2
pyobjc-framework-MediaToolbox==9.2
pyobjc-framework-Metal==9.2
pyobjc-framework-MetalFX==9.2
pyobjc-framework-MetalKit==9.2
pyobjc-framework-MetalPerformanceShaders==9.2
pyobjc-framework-MetalPerformanceShadersGraph==9.2
pyobjc-framework-MetricKit==9.2
pyobjc-framework-MLCompute==9.2
pyobjc-framework-ModelIO==9.2
pyobjc-framework-MultipeerConnectivity==9.2
pyobjc-framework-NaturalLanguage==9.2
pyobjc-framework-NetFS==9.2
pyobjc-framework-Network==9.2
pyobjc-framework-NetworkExtension==9.2
pyobjc-framework-NotificationCenter==9.2
pyobjc-framework-OpenDirectory==9.2
pyobjc-framework-OSAKit==9.2
pyobjc-framework-OSLog==9.2
pyobjc-framework-PassKit==9.2
pyobjc-framework-PencilKit==9.2
pyobjc-framework-PHASE==9.2
pyobjc-framework-Photos==9.2
pyobjc-framework-PhotosUI==9.2
pyobjc-framework-PreferencePanes==9.2
pyobjc-framework-PushKit==9.2
pyobjc-framework-Quartz==9.2
pyobjc-framework-QuickLookThumbnailing==9.2
pyobjc-framework-ReplayKit==9.2
pyobjc-framework-SafariServices==9.2
pyobjc-framework-SafetyKit==9.2
pyobjc-framework-SceneKit==9.2
pyobjc-framework-ScreenCaptureKit==9.2
pyobjc-framework-ScreenSaver==9.2
pyobjc-framework-ScreenTime==9.2
pyobjc-framework-ScriptingBridge==9.2
pyobjc-framework-SearchKit==9.2
pyobjc-framework-Security==9.2
pyobjc-framework-SecurityFoundation==9.2
pyobjc-framework-SecurityInterface==9.2
pyobjc-framework-ServiceManagement==9.2
pyobjc-framework-SharedWithYou==9.2
pyobjc-framework-SharedWithYouCore==9.2
pyobjc-framework-ShazamKit==9.2
pyobjc-framework-Social==9.2
pyobjc-framework-SoundAnalysis==9.2
pyobjc-framework-Speech==9.2
pyobjc-framework-SpriteKit==9.2
pyobjc-framework-StoreKit==9.2
pyobjc-framework-SyncServices==9.2
pyobjc-framework-SystemConfiguration==9.2
pyobjc-framework-SystemExtensions==9.2
pyobjc-framework-ThreadNetwork==9.2
pyobjc-framework-UniformTypeIdentifiers==9.2
pyobjc-framework-UserNotifications==9.2
pyobjc-framework-UserNotificationsUI==9.2
pyobjc-framework-VideoSubscriberAccount==9.2
pyobjc-framework-VideoToolbox==9.2
pyobjc-framework-Virtualization==9.2
pyobjc-framework-Vision==9.2
pyobjc-framework-WebKit==9.2
pyparsing==3.0.9
PyPDF2==3.0.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.2
scipy==1.11.1
seaborn==0.12.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.138
uri-template==1.3.0
urllib3==2.0.4
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
widgetsnbextension==4.0.8
